database/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
google-cloud-secret-manager
google_cloud-sql-python-connector
google-cloud-storage
